move repl web component to new repl package + use it in /vanilla + /vanilla/mini
