+++ packages/repl/README.md
@@ -1,0 +1,3 @@
+# @strudel/repl
+
+The Strudel REPL as a web component.
