+++ packages/repl/index.mjs
@@ -1,0 +1,1 @@
+export * from './repl-component.mjs';
