+++ website/src/pages/vanilla/index.astro
@@ -90,7 +90,8 @@ import HeadCommonNew from '../../components/HeadCommonNew.astro';
         <!-- <label><input type="checkbox" name="isTooltipEnabled" />isTooltipEnabled</label> -->
       </form>
     </div>
-    <div id="code"></div>
+    <strudel-editor id="editor"></strudel-editor>
     <script src="../../repl/vanilla/vanilla.mjs"></script>
+    <script src="@strudel/repl"></script>
   </body>
 </html>
